Conversion to Typescript
