+++ app/page.tsx
@@ -2,7 +2,5 @@ import Image from "next/image";
 import App from "@components/App";
 
 export default function Home() {
-  return (
-    <App/>
-  );
+  return <App />;
 }
