+++ components/HowToBuy.tsx
