+++ components/Tokenomics.tsx
