+++ package.json
@@ -15,7 +15,10 @@
     "react-dom": "^18"
   },
   "devDependencies": {
+    "@types/node": "^20.14.2",
+    "@types/react": "^18.3.3",
     "postcss": "^8",
-    "tailwindcss": "^3.4.1"
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.4.5"
   }
 }
